fixed in role permission

+++ app/Http/Controllers/Api/MaterialController.php
@@ -9,16 +9,13 @@ use App\Http\Resources\MaterialResource;
 use App\Models\Material;
 use App\Services\MaterialService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class MaterialController extends Controller
 {
   public function __construct(
     private MaterialService $materialService
-  ) {
-    $this->authorizeResource(Material::class, 'material', [
-      'except' => ['index']
-    ]);
-  }
+  ) {}
 
   public function index(Request $request)
   {
@@ -33,22 +30,29 @@ class MaterialController extends Controller
 
   public function store(CreateMaterialRequest $request)
   {
+    Gate::authorize('create', Material::class);
+
     $material = $this->materialService->createMaterial($request->validated());
     return new MaterialResource($material);
   }
 
   public function show(Material $material)
   {
+
     return $material;
   }
 
   public function update(Material $material, UpdateMaterialRequest $request)
   {
+    Gate::authorize('update', $material);
+
     $newMaterial = $this->materialService->updateMaterial($material, $request->validated());
     return new MaterialResource($newMaterial);
   }
   public function destroy(Material $material)
   {
+    Gate::authorize('delete', $material);
+
     $material = $this->materialService->deleteMaterial($material);
     return response()->json(['message' => 'Material deleted successfully']);
   }
